Rename Knight script
